Made accesses to ApplicationIconBadgeNumber always use the main queue.
In Xamarin.iOS, any accessess made to a UIKit property that's labeled as `nonatomic` on a thread other than the main thread results in an exception being thrown. This was causing installation saves to fail quietly (very bad).

We solve this problem by forcing usage of these properties onto the main queue, and although this provides an avenue for deadlock, it's better than simply having silent failures in most cases.