Made all methods in ReflectionHelpers non-extensions.
Reason: as we start adding support for more platforms which may or may not have this natively, our extension methods are not guaranteed to be invoked if they have the same name of a builtin method.

My requiring explicit invoking, it allows us to enforce custom logic here.

Fixes #111 (for real this time).