Made TimeZone string generation locale-agnostic for Windows Phone and Unity.
Previously, TimeZone string generation that we sent along with
ParseInstallation was dependent upon the CurrentCulture of the device
being set to the english locale (as StandardName is localized, WHAT?!)

Our new, fancy TimeZone string is now generated agnostically of the
current device locale. There is potential for information about the
specific region of the time zone to be lost here, but most applications
should not require this.

This should fix #138.