Refactor ConvertToJSON method in IJsonConvertible classes
Updated the `ConvertToJSON` method in multiple classes implementing the `IJsonConvertible` interface to return an `IDictionary<string, object>` instead of an `object`. This change improves the method's capability to generate structured JSON-compatible representations, accommodating more complex data structures. A new `ConvertValueToJSON` method was also added to the `ParseSetOperation` class to assist in converting values to JSON format.